huge update - don't use individual axes, but do everything on one axis
This basically required changing all the plotting code. The code is still a bit sloppy now, I'll clean it up and update it later.

This update makes the code run much faster. It's way easier to generate many tables, as they can be programatically updated without making a whole new set of axes